requirements.txt
joblib==1.3.2
numpy==1.25.2
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.2
sklearn==0.0.post7
threadpoolctl==3.2.0
